+++ src/User/RegisterUserCommand.php
@@ -49,6 +49,12 @@ final class RegisterUserCommand
     /** @var string|null */
     private $intraEuropeanCommunityVAT;
 
+    /** @var string|null */
+    private $company;
+
+    /** @var string|null */
+    private $function;
+
     /** @return string|null */
     public function getPhone(): ?string
     {
@@ -187,9 +193,7 @@ final class RegisterUserCommand
         return $this;
     }
 
-    /**
-     * @return UpdateUserAddressCommand|null
-     */
+    /** @return UpdateUserAddressCommand|null */
     public function getBilling(): ?UpdateUserAddressCommand
     {
         return $this->billing;
@@ -207,9 +211,7 @@ final class RegisterUserCommand
         return $this;
     }
 
-    /**
-     * @return UpdateUserAddressCommand|null
-     */
+    /** @return UpdateUserAddressCommand|null */
     public function getShipping(): ?UpdateUserAddressCommand
     {
         return $this->shipping;
@@ -227,11 +229,17 @@ final class RegisterUserCommand
         return $this;
     }
 
+    /** @return string|null */
     public function getExternalIdentifier(): ?string
     {
         return $this->externalIdentifier;
     }
 
+    /**
+     * @param string|null $externalIdentifier
+     *
+     * @return $this
+     */
     public function setExternalIdentifier(?string $externalIdentifier): self
     {
         $this->externalIdentifier = $externalIdentifier;
@@ -239,11 +247,19 @@ final class RegisterUserCommand
         return $this;
     }
 
+
+    /** @return bool|null */
     public function getIsProfessional(): ?bool
     {
         return $this->isProfessional;
     }
 
+
+    /**
+     * @param bool|null $isProfessional
+     *
+     * @return $this
+     */
     public function setIsProfessional(?bool $isProfessional): self
     {
         $this->isProfessional = $isProfessional;
@@ -251,14 +267,56 @@ final class RegisterUserCommand
         return $this;
     }
 
+    /** @return string|null */
     public function getIntraEuropeanCommunityVAT(): ?string
     {
         return $this->intraEuropeanCommunityVAT;
     }
 
+    /**
+     * @param string|null $intraEuropeanCommunityVAT
+     *
+     * @return $this
+     */
     public function setIntraEuropeanCommunityVAT(?string $intraEuropeanCommunityVAT): self
     {
         $this->intraEuropeanCommunityVAT = $intraEuropeanCommunityVAT;
+
+        return $this;
+    }
+
+    /** @return string|null */
+    public function getCompany(): ?string
+    {
+        return $this->company;
+    }
+
+    /**
+     * @param string|null $company
+     *
+     * @return $this
+     */
+    public function setCompany(?string $company): self
+    {
+        $this->company = $company;
+
+        return $this;
+    }
+
+    /** @return string|null */
+    public function getFunction(): ?string
+    {
+        return $this->function;
+    }
+
+    /**
+     * @param string|null $function
+     *
+     * @return $this
+     */
+    public function setFunction(?string $function): self
+    {
+        $this->function = $function;
 
         return $this;
     }